Split behavior method collection tests
